✨ feat(StreamKey.Core): Add query authentication to HttpClient
Adds a new `AddQueryAuth` extension method to `IQueryCollection` to apply authorization and device ID headers from query parameters to an `HttpClient`. This refactors the TwitchService to use this new method, simplifying header management. Also removes the use of `IHttpClientFactory` and directly injects `HttpClient` for TwitchService, along with related configuration changes.